requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==5.1.1
djangorestframework==3.13.1
idna==3.3
py==1.11.0
pytz==2022.1
requests==2.27.1
retry==0.9.2
sqlparse==0.4.2
urllib3==1.26.9
